creating a simple page using LinearLayout
LinearLayout is a view Group. It is used when we have a given component and we want to view them either horizontally or vertically depending upon the orientation we want.
Here, rest of coding like creating a main_activity is the same only we have to do this coding in activity_main.xml.